AI_Services_Flask_App/requirements.txt
Flask==2.2.2
Werkzeug==2.2.3
flask-cors==4.0.0
python-dotenv==1.0.0
chromadb==0.3.21
numpy==1.23.5
scikit-learn==1.1.3
pandas==1.5.3
gunicorn==20.1.0
flasgger==0.9.7.1
PyPDF2==3.0.1
requests==2.31.0
elevenlabs==1.50.0